Separating WiFiConnectionManager definition/implementation in a h/cpp file
