+++ src/WiFiConnectionManager.h
@@ -15,15 +15,23 @@
  * a commercial license, send an email to license@arduino.cc.
  */
 
+/******************************************************************************
+ * INCLUDE
+ ******************************************************************************/
+
 #include "ConnectionManager.h"
 #include <WiFiUdp.h>
+
+/******************************************************************************
+ * CLASS DECLARATION
+ ******************************************************************************/
 
 class WiFiConnectionManager : public ConnectionManager {
 public:
   WiFiConnectionManager(const char *ssid, const char *pass);
 
-  virtual unsigned long getTime();
   virtual void init();
+  virtual unsigned long getTime();
   virtual void check();
   virtual Client &getClient() { return wifiClient; };
   virtual UDP &getUDP() { return udp; };
@@ -52,138 +60,3 @@ private:
   WiFiClient wifiClient;
   int connectionTickTimeInterval;
 };
-
-static const unsigned long NETWORK_CONNECTION_INTERVAL = 30000;
-
-WiFiConnectionManager::WiFiConnectionManager(const char *ssid, const char *pass) :
-  ssid(ssid), pass(pass),
-  lastConnectionTickTime(millis()),
-  connectionTickTimeInterval(CHECK_INTERVAL_IDLE),
-  getTimeRetries(MAX_GETTIME_RETRIES) {
-}
-
-unsigned long WiFiConnectionManager::getTime() {
-  return WiFi.getTime();
-}
-
-void WiFiConnectionManager::init() {
-}
-
-void WiFiConnectionManager::changeConnectionState(NetworkConnectionState _newState) {
-  char msgBuffer[120];
-  int newInterval = CHECK_INTERVAL_INIT;
-  switch (_newState) {
-    case CONNECTION_STATE_INIT:
-      newInterval = CHECK_INTERVAL_INIT;
-      break;
-    case CONNECTION_STATE_CONNECTING:
-      sprintf(msgBuffer, "Connecting to \"%s\"", ssid);
-      debugMessage(msgBuffer, 2);
-      newInterval = CHECK_INTERVAL_CONNECTING;
-      break;
-    case CONNECTION_STATE_GETTIME:
-      newInterval = CHECK_INTERVAL_GETTIME;
-      debugMessage("Acquiring Time from Network", 3);
-      getTimeRetries = MAX_GETTIME_RETRIES;
-      break;
-    case CONNECTION_STATE_CONNECTED:
-      newInterval = CHECK_INTERVAL_CONNECTED;
-      break;
-    case CONNECTION_STATE_DISCONNECTED:
-      sprintf(msgBuffer, "WiFi.status(): %d", WiFi.status());
-      debugMessage(msgBuffer, 4);
-      sprintf(msgBuffer, "Connection to \"%s\" lost.", ssid);
-      debugMessage(msgBuffer, 0);
-      debugMessage("Attempting reconnection", 0);
-      newInterval = CHECK_INTERVAL_DISCONNECTED;
-      break;
-    case CONNECTION_STATE_ERROR:
-      debugMessage("WiFi Hardware failure.\nMake sure you are using a WiFi enabled board/shield.", 0);
-      debugMessage("Then reset and retry.", 0);
-      break;
-  }
-  connectionTickTimeInterval = newInterval;
-  lastConnectionTickTime = millis();
-  netConnectionState = _newState;
-}
-
-void WiFiConnectionManager::check() {
-  char msgBuffer[120];
-  unsigned long const now = millis();
-  int networkStatus = 0;
-  if (now - lastConnectionTickTime > connectionTickTimeInterval) {
-    switch (netConnectionState) {
-      case CONNECTION_STATE_INIT:
-        networkStatus = WiFi.status();
-        sprintf(msgBuffer, "WiFi.status(): %d", networkStatus);
-        debugMessage(msgBuffer, 2);
-        if (networkStatus == NETWORK_HARDWARE_ERROR) {
-          // NO FURTHER ACTION WILL FOLLOW THIS
-          changeConnectionState(CONNECTION_STATE_ERROR);
-          lastConnectionTickTime = now;
-          return;
-        }
-        sprintf(msgBuffer, "Current WiFi Firmware: %s", WiFi.firmwareVersion());
-        debugMessage(msgBuffer, 0);
-        if(WiFi.firmwareVersion() < WIFI_FIRMWARE_VERSION_REQUIRED){
-          sprintf(msgBuffer, "Latest WiFi Firmware: %s", WIFI_FIRMWARE_VERSION_REQUIRED);
-          debugMessage(msgBuffer, 0);
-          debugMessage("Please update to the latest version for best performance.", 0);
-          delay(5000);
-        }
-        changeConnectionState(CONNECTION_STATE_CONNECTING);
-        break;
-      case CONNECTION_STATE_CONNECTING:
-        networkStatus = WiFi.begin(ssid, pass);
-        sprintf(msgBuffer, "WiFi.status(): %d", networkStatus);
-        debugMessage(msgBuffer, 4);
-        if (networkStatus != NETWORK_CONNECTED) {
-          sprintf(msgBuffer, "Connection to \"%s\" failed", ssid);
-          debugMessage(msgBuffer, 0);
-          sprintf(msgBuffer, "Retrying in  \"%d\" milliseconds", connectionTickTimeInterval);
-          debugMessage(msgBuffer, 2);
-          //changeConnectionState(CONNECTION_STATE_CONNECTING);
-          return;
-        } else {
-          sprintf(msgBuffer, "Connected to \"%s\"", ssid);
-          debugMessage(msgBuffer, 2);
-          changeConnectionState(CONNECTION_STATE_GETTIME);
-          return;
-        }
-        break;
-      case CONNECTION_STATE_GETTIME:
-        unsigned long networkTime;
-        networkTime = getTime();
-        debugMessage(".", 3, false, false);
-        if(networkTime > lastValidTimestamp){
-          debugMessage("", 3, false, true);
-          lastValidTimestamp = networkTime;
-          sprintf(msgBuffer, "Network Time: %u", networkTime);
-          debugMessage(msgBuffer, 3);
-          changeConnectionState(CONNECTION_STATE_CONNECTED);
-        } else if (WiFi.status() != WL_CONNECTED) {
-           changeConnectionState(CONNECTION_STATE_DISCONNECTED);
-        } else if (!getTimeRetries--) {
-           changeConnectionState(CONNECTION_STATE_DISCONNECTED);
-        } 
-        break;
-      case CONNECTION_STATE_CONNECTED:
-        // keep testing connection
-        networkStatus = WiFi.status();
-        sprintf(msgBuffer, "WiFi.status(): %d", networkStatus);
-        debugMessage(msgBuffer, 4);
-        if (networkStatus != WL_CONNECTED) {
-          changeConnectionState(CONNECTION_STATE_DISCONNECTED);
-          return;
-        }
-        sprintf(msgBuffer, "Connected to \"%s\"", ssid);
-        debugMessage(msgBuffer, 4);
-        break;
-      case CONNECTION_STATE_DISCONNECTED:
-        WiFi.end();
-        changeConnectionState(CONNECTION_STATE_CONNECTING);
-        break;
-    }
-    lastConnectionTickTime = now;
-  }
-}
